+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to the "claude-dev" extension will be documented in this fil
 
 ## [1.1.1]
 
-- Added the ability to choose other Claude models (+ GPT-4o, Gemini, DeepSeek, and Mistral if you use OpenRouter)
+- Added the ability to choose other Claude models (+ GPT-4o, DeepSeek, and Mistral if you use OpenRouter)
 
 ## [1.1.0]